+++ packages/terra-core-docs/src/terra-dev-site/doc/show-hide/example/InitiallyOpenShowHide.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
-import ShowHide from 'terra-show-hide';
+import ShowHide, { ShowHideFocuser } from 'terra-show-hide';
 
-const fullText = 'Cerner cardiovascular solutions are embedded within the EHR, allowing for a holistic patient record that includes: diagnostic activities, therapeutic interventions and follow-up regimens. Electrocardiogram (ECG) objects are digitized and accessed within the EHR providing physicians in your cardiology department the ability to receive and present ECG data in mere seconds after it is acquired. This eliminates the necessity for a separate ECG information system.';
-const previewText = fullText.substring(0, 280);
+const prefix = 'Cerner cardiovascular solutions are embedded within the EHR, allowing for a holistic patient record that includes: diagnostic activities, therapeutic interventions and follow-up regimens. Electrocardiogram (ECG) objects are digitized and accessed within the EHR providing physicians';
+const focusableText = 'in your cardiology department the ability to receive and present ECG data in mere seconds after it is acquired. This eliminates the necessity for a separate ECG information system.';
 
 class InitiallyOpenShowHide extends React.Component {
   constructor(props) {
@@ -10,6 +10,7 @@ class InitiallyOpenShowHide extends React.Component {
 
     this.state = { isOpen: true };
     this.toggleShowHide = this.toggleShowHide.bind(this);
+    this.focusRef = React.createRef();
   }
 
   toggleShowHide() {
@@ -20,8 +21,10 @@ class InitiallyOpenShowHide extends React.Component {
 
   render() {
     return (
-      <ShowHide preview={<p>{previewText}</p>} isOpen={this.state.isOpen} onChange={this.toggleShowHide}>
-        <p>{fullText}</p>
+      <ShowHide focusRef={this.focusRef} preview={<p>{prefix}</p>} isOpen={this.state.isOpen} onChange={this.toggleShowHide}>
+        <p>
+          <ShowHideFocuser ref={this.focusRef} prefix={prefix} focusableText={focusableText} />
+        </p>
       </ShowHide>
     );
   }