+++ packages/terra-core-docs/CHANGELOG.md
@@ -5,6 +5,13 @@
 * Changed
   * Updated `terra-button-group` tests.
 
+* Added
+  * Added Search field examples for optional Label with default and custom labels.
+
+* Changed
+  * Updated default search delay to 2500ms.
+  * Update Search field examples to be more functionality focused. 
+
 ## 1.36.0 - (August 11, 2023)
 
 * Changed
@@ -40,6 +47,7 @@
 * Updated
   * Updated an example for `terra-dropdown-button`.
   * Updated email field validation for `terra-form-field`.
+  * Updated `terra-show-hide` examples for new focusRef prop, description for the new prop usage.
 
 * Added
   * Added documentation updates for `terra-form-input`.